FOUR-11898 Saved Filter - Requests - Save in user session
