+++ resources/js/tasks/components/TasksList.vue
@@ -1,14 +1,7 @@
 <template>
   <div class="data-table">
-    <data-loading
-      v-show="shouldShowLoader"
-      :for="/tasks\?page|results\?page/"
-      :empty="$t('Congratulations')"
-      :empty-desc="$t('You don\'t currently have any tasks assigned to you')"
-      empty-icon="beach"
-    />
     <div
-      v-show="!shouldShowLoader"
+      v-show="true"
       data-cy="tasks-table"
     >
       <filter-table
@@ -89,6 +82,13 @@
           </td>
         </template>
       </filter-table>
+      <data-loading
+        v-show="shouldShowLoader"
+        :for="/tasks\?page|results\?page/"
+        :empty="$t('Congratulations')"
+        :empty-desc="$t('You don\'t currently have any tasks assigned to you')"
+        empty-icon="beach"
+      />
       <pagination-table
         :meta="data.meta"
         @page-change="changePage"
@@ -110,6 +110,7 @@ import { FilterTable } from "../../components/shared";
 import TasksPreview from "./TasksPreview.vue";
 import ListMixin from "./ListMixin";
 import PMColumnFilterPopover from "../../components/PMColumnFilterPopover/PMColumnFilterPopover.vue";
+import PMColumnFilterPopoverTasksMixin from "./PMColumnFilterPopoverTasksMixin.js";
 import paginationTable from "../../components/shared/PaginationTable.vue";
 
 const uniqIdsMixin = createUniqIdsMixin();
@@ -123,7 +124,7 @@ export default {
     PMColumnFilterPopover,
     paginationTable,
   },
-  mixins: [datatableMixin, dataLoadingMixin, uniqIdsMixin, ListMixin],
+  mixins: [datatableMixin, dataLoadingMixin, uniqIdsMixin, ListMixin, PMColumnFilterPopoverTasksMixin],
   props: {
     filter: {},
     columns: {},
@@ -152,7 +153,6 @@ export default {
       ],
       orderBy: "ID",
       order_direction: "DESC",
-      advancedFilter: [],
       status: "",
       sortOrder: [
         {
@@ -219,7 +219,7 @@ export default {
           sortable: true,
           default: true,
           width: 140,
-          truncate: true,
+          truncate: true
         },
         {
           label: "REQUEST",
@@ -349,115 +349,6 @@ export default {
       this.page = page;
       this.fetch();
     },
-    onChangeSort(value) {
-      this.order_direction = value;
-    },
-    onApply(json, index) {
-      this.advancedFilter[index] = json;
-      console.log(this.advancedFilter, index);
-      this.fetch();
-    },
-    onClear(index) {
-      this.advancedFilter[index] = [];
-      this.fetch();
-    },
-    onUpdate(object, index) {
-      if (object.$refs.pmColumnFilterForm && 
-        this.advancedFilter.length > 0 &&
-        this.advancedFilter[index] && 
-        this.advancedFilter[index].length > 0) {
-        object.$refs.pmColumnFilterForm.setValues(this.advancedFilter[index]);
-      }
-    },
-    getAdvancedFilter() {
-      let flat = this.advancedFilter.flat(1);
-      return flat.length > 0 ? "&advanced_filter=" + JSON.stringify(flat) : "";
-    },
-    getFormat(column) {
-      let format = "string";
-      if (column.format) {
-        format = column.format;
-      }
-      if (column.field === "status" || column.field === "assignee") {
-        format = "stringSelect";
-      }
-      return format;
-    },
-    getFormatRange(column) {
-      let formatRange = [];
-      if (column.field === "status") {
-        formatRange = ["In Progress", "Completed", "Error", "Canceled"];
-      }
-      if (column.field === "assignee") {
-        formatRange = ["user1", "user2", "user3", "user4"];
-      }
-      return formatRange;
-    },
-    getOperators(column) {
-      let operators = [];
-      if (column.field === "status" || column.field === "assignee") {
-        operators = ["=", "in"];
-      }
-      return operators;
-    },
-    getViewConfigFilter() {
-      return [
-        {
-          "type": "string",
-          "includes": ["=", "<", "<=", ">", ">=", "contains", "regex"],
-          "control": "PMColumnFilterOpInput",
-          "input": ""
-        },
-        {
-          "type": "string",
-          "includes": ["between"],
-          "control": "PMColumnFilterOpBetween",
-          "input": []
-        },
-        {
-          "type": "string",
-          "includes": ["in"],
-          "control": "PMColumnFilterOpIn",
-          "input": []
-        },
-        {
-          "type": "datetime",
-          "includes": ["=", "<", "<=", ">", ">=", "contains", "regex"],
-          "control": "PMColumnFilterOpDatetime",
-          "input": ""
-        },
-        {
-          "type": "datetime",
-          "includes": ["between"],
-          "control": "PMColumnFilterOpBetweenDatepicker",
-          "input": []
-        },
-        {
-          "type": "datetime",
-          "includes": ["in"],
-          "control": "PMColumnFilterOpInDatepicker",
-          "input": []
-        },
-        {
-          "type": "stringSelect",
-          "includes": ["="],
-          "control": "PMColumnFilterOpSelect",
-          "input": ""
-        },
-        {
-          "type": "stringSelect",
-          "includes": ["in"],
-          "control": "PMColumnFilterOpSelectMultiple",
-          "input": []
-        },
-        {
-          "type": "boolean",
-          "includes": ["="],
-          "control": "PMColumnFilterOpBoolean",
-          "input": false
-        }
-      ];
-    }
   },
 };
 </script>
